# Fix interpolation extremities on exponential and elastic curves