Resolves #3015.

# Summary

`Elastic` and `Expo` curves due to the formulae used would not stop exactly at 0 or 1 at time 0 or 1, as pointed out in tests in c84b164. Resolve this discrepancy by introducing small linear factors to the appropriate easing curves.

The linear factors offset the differences at the extremities that were wrong, and then fade to 0 at the other ends which were correct. A special case are the `InOut` easings, in which the factor will be fully included at both ends and then fade out linearly to 0 at `time = 0.5`.

Since the adjustments involved are minuscule (less than 4 significant decimal places), they should not be noticeable during transitions, but fix the jumping at extreme points seen before.

# Remarks

Aside from the unit tests originally written by @peppy (sorry for taking co-credit on that commit, I don't quite know why that happened - might have been due to me rebasing a bit to clean up history before push) I also included a test scene that allows for previewing all available curves for comparison reasons.

To check the visual effects you can either run the test scene or see the attached videos: [before](https://streamable.com/94t0e), [after](https://streamable.com/ct70d). The difference is quite subtle and will mostly be apparent on large sizes, and at the very ends of the transitions.