@Tom94 2019-11-25
I'm not certain how good the JIT optimizations are, but it might make sense to bake all the constants into static variables to make sure they're not re-computed all the time. Pow is relatively expensive, after all. Perhaps worth a test.