@bdach 2019-11-25
Yeah, I figured it might be worth it. Will benchmark and report back.