@bdach 2019-11-25
Benchmarking results ([see benchmark code](https://gist.github.com/bdach/0edda878ee2664faf2bbd0056fdfe9ca)):

### Before extractions to constants

|      Method |            Easing |     Mean |   Error |   StdDev |
|------------ |------------------ |---------:|--------:|---------:|
| ApplyEasing |            InExpo | 101.4 us | 2.00 us |  3.71 us |
| ApplyEasing |         InElastic | 114.5 us | 2.19 us |  2.34 us |
| ApplyEasing |    OutElasticHalf | 128.5 us | 2.55 us |  4.05 us |
| ApplyEasing | OutElasticQuarter | 127.6 us | 2.55 us |  3.22 us |
| ApplyEasing |      InOutElastic | 134.7 us | 3.95 us | 11.65 us |

### After extractions to constants

|      Method |            Easing |     Mean |    Error |   StdDev |
|------------ |------------------ |---------:|---------:|---------:|
| ApplyEasing |            InExpo | 46.87 us | 0.635 us | 0.594 us |
| ApplyEasing |         InElastic | 64.19 us | 1.146 us | 0.957 us |
| ApplyEasing |    OutElasticHalf | 62.81 us | 0.950 us | 0.889 us |
| ApplyEasing | OutElasticQuarter | 59.69 us | 0.533 us | 0.499 us |
| ApplyEasing |      InOutElastic | 65.64 us | 0.658 us | 0.616 us |

Yup, seems like a definite win performance-wise; just pushed that in.